+++ ui/src/components/forms/profileForm/documentUpload/DocumentUpload.js
@@ -1,11 +1,8 @@
 import React, { useState } from 'react';
 import './documentUpload.css';
-import { ButtonWithCancel } from '../../../basic/button/Button';
-import Dropzone from './Dropzone';
+import UploadModal from './UploadModal';
 import DropdownSelector from '../../../basic/dropdownSelector/DropdownSelector';
-import passportDoc from '../../../../assets/exampleImageSeeds/passport.json';
-import driversLicenseDocs from '../../../../assets/exampleImageSeeds/driversLicense.json';
-import governmentIdDocs from '../../../../assets/exampleImageSeeds/governmentId.json';
+import Input from '../../../basic/input/Input';
 
 
 // example document structure 
@@ -22,13 +19,13 @@ import governmentIdDocs from '../../../../assets/exampleImageSeeds/governmentId.
 //   ],
 // };
 
-const DocumentTypes = {
+export const DocumentTypes = {
   PASSPORT: 'passport',
   DRIVERS_LICENSE: 'driversLicense',
   GOVERNMENT_ID: 'governmentId'
 }
 
-const photoIdLabels = {
+export const photoIdLabels = {
   [DocumentTypes.PASSPORT]: 'Passport',
   [DocumentTypes.DRIVERS_LICENSE]: `Driver's License`,
   [DocumentTypes.GOVERNMENT_ID]: 'Government Issued ID'
@@ -43,9 +40,10 @@ const photoIdOptions = [
 const dropdownPlaceholder = 'Choose ID Type';
 
 
-const DocumentUpload = ({ isUploaded, onUpload, onClear }) => {
+const DocumentUpload = ({ onUpload, onClear, documents }) => {
   const [isUploading, setIsUploading] = useState(false);
   const [documentType, setDocumentType] = useState(dropdownPlaceholder);
+  const documentNames = documents && documents.files.map(d => d.description).join(', ');
   
   const uploadToProfileForm = files => onUpload({
     files,
@@ -56,119 +54,41 @@ const DocumentUpload = ({ isUploaded, onUpload, onClear }) => {
   return (
     <div>
       {
-        isUploaded ?
-        <p><span className="OnrampColor BoldText Pointer" onClick={onClear}>clear</span></p>
+        Boolean(documents) ?
+        <div className="Relative">
+          <Input className="DocumentImageInput" value={documentNames} placeholder="Photo ID" readOnly={true} />
+          <RemoveEdit onRemove={onClear} onEdit={() => setIsUploading(true)} />
+        </div>
         :
-        <>
-          <DropdownSelector
-            selectedOption={documentType}
-            selectOption={option => {
-              setDocumentType(option);
-              setIsUploading(true);
-            }}
-            options={photoIdOptions}
-            selectedClassName={documentType === dropdownPlaceholder ? 'LightGreyText ThinText MediumTextSize' : 'BlackText ThinText MediumTextSize'}
-          />
-          {
-            isUploading &&
-            <UploadModal
-              documentType={documentType}
-              onUpload={uploadToProfileForm}
-              closeModal={() => setIsUploading(false)}
-              clearDocumentType={() => setDocumentType(dropdownPlaceholder)}
-            />
-          }
-        </>
+        <DropdownSelector
+          selectedOption={documentType}
+          selectOption={option => {
+            setDocumentType(option);
+            setIsUploading(true);
+          }}
+          options={photoIdOptions}
+          selectedClassName={documentType === dropdownPlaceholder ? 'LightGreyText ThinText MediumTextSize' : 'BlackText ThinText MediumTextSize'}
+        />
       }
+      {
+        isUploading &&
+        <UploadModal
+          documentsOnForm={documents}
+          documentType={documentType === dropdownPlaceholder ? DocumentTypes.PASSPORT : documentType}
+          onUpload={uploadToProfileForm}
+          closeModal={() => setIsUploading(false)}
+          clearDocumentType={() => setDocumentType(dropdownPlaceholder)}
+        />
+          }
     </div>
   )
 }
 
-const documentTypeDescriptions = {
-  passport: 'Please provide the first page of your passport with your photo on it.',
-  driversLicense: `Please provide the front and back of your driver's license.`,
-  governmentId: `Please provide the front and back of your government issued ID.`,
-  shared: 'Make sure the information on the image is clear and legible. Accepted file formats are: png, jpg, pdf, doc, docx.'
-}
-
-const UploadModal = ({ closeModal, onUpload, documentType, clearDocumentType }) => {
-  const photoIdLabel = photoIdLabels[documentType];
-  const twoDropzones = documentType !== DocumentTypes.PASSPORT;
-  const dropzoneLabels = twoDropzones ? [`Front of ${photoIdLabel}`, `Back of ${photoIdLabel}`] : ['',''];
-  const initialDocumentArray = twoDropzones ? ['','']: [''];
-  const [documents, loadDocuments] = useState(initialDocumentArray);
-
-  const onDrop = index => acceptedFiles => {
-    acceptedFiles.forEach((file, j) => {
-      if (j > 0) return;
-      const reader = new FileReader();
-      reader.onload = e => {
-        const newDocument = {
-          data: e.target.result,
-          media_type: file.type,
-          description: file.name,
-          orientation: index === 0 ? 'front' : 'back',
-        };
-        if (index === 0) {
-          const newDocuments = twoDropzones ? [newDocument, documents[1]] : [newDocument];
-          loadDocuments(newDocuments);
-        } else {
-          loadDocuments([documents[0], newDocument]);
-        }
-      };
-      reader.readAsDataURL(file);
-    });
-  };
-
-  const onClearImage = index => () => {
-    if (index === 0) {
-      loadDocuments(['', documents[1]]);
-    } else {
-      loadDocuments([documents[0], '']);
-    }
-  }
-
-  const loadExampleImage = () => {
-    if (documentType === DocumentTypes.PASSPORT) { return loadDocuments(passportDoc) }
-    if (documentType === DocumentTypes.DRIVERS_LICENSE) { return loadDocuments(driversLicenseDocs) }
-    if (documentType === DocumentTypes.GOVERNMENT_ID) { return loadDocuments(governmentIdDocs) }
-  }
-
-  const cancelModal = () => {
-    closeModal();
-    clearDocumentType();
-  }
-
-  const filesAreUploaded = documents.reduce((acc, doc) => acc && Boolean(doc), true);
-
-  return (
-    <div className="UploadModal">
-      <div className="UploadModalBackground" onClick={cancelModal} />
-      <div className="ModalBox">
-        <div className="ModalBoxHeader">
-          <h2 className="ModalTitle">{`Upload ${photoIdLabel}`}</h2>
-        </div>
-        <p className="DocumentTypeDescription">{`${documentTypeDescriptions[documentType]} ${documentTypeDescriptions.shared}`}</p>
-        <p className="OnrampColor ExtraBold Pointer" onClick={loadExampleImage}>Use example image</p>
-        <div className={`DropzoneContainer ${twoDropzones ? 'DropzoneGrid' : ''}`}>
-          <Dropzone onDrop={onDrop(0)} onClear={onClearImage(0)} document={documents[0]} label={dropzoneLabels[0]} />
-          {twoDropzones && <Dropzone onDrop={onDrop(1)} onClear={onClearImage(1)} document={documents[1]} label={dropzoneLabels[1]} />}
-        </div>
-        <div className="ModalButtonContainer">
-          <ButtonWithCancel
-            primaryColor="#9C27B0"
-            disabled={!filesAreUploaded}
-            onClick={() => {
-              console.log(JSON.stringify(documents))
-              onUpload(documents);
-              closeModal();
-            }}
-            onCancel={cancelModal}
-          >Save</ButtonWithCancel>
-        </div>
-      </div>
-    </div>
-  )
-}
+const RemoveEdit = ({ onRemove, onEdit }) => (
+  <div className="RemoveEditContainer">
+    <p className="Remove" onClick={onRemove}>Remove</p>
+    <p className="Edit" onClick={onEdit}>Edit</p>
+  </div>
+)
 
 export default DocumentUpload;
